\name{avgTestAcc}
\alias{avgTestAcc}
\title{
Average test accuracy for a classification ensemble
}
\description{
The avgTestAcc function returns the average test accuracy of an ensemble of classifiers as generated by the function \code{\link{ensRBF}}.
}
\usage{
avgTestAcc(ens)
}
\arguments{
  \item{ens}{The R object as generated by \code{\link{ensRBF}}}
}
\value{
Returns a numerical value, equal to the average overall test accuracy (\%CC) of the ensemble. 
}
\references{
  \url{http://www.classyfire.org}
}
\seealso{
  \code{\link{testAcc}},
  \code{\link{confMatr}}
}
\author{
Eleni Chatzimichali <hatzimihali.eleni@gmail.com>,
Conrad Bessant <c.bessant@qmul.ac.uk>
}
\examples{

data(iris)

sampleIris = iris[,5]
iris = iris[,-5]
            
obj <- ensRBF(iris, sampleIris, bootNum = 10, ensembleNum = 20, runParallel = TRUE, 
              cpus = 4, type = "SOCK")

# Return the average test accuracy for the ensemble

avgTestAcc(obj)

}
