\name{avgTrainAcc}
\alias{avgTrainAcc}
\title{
Average train accuracy for a classification ensemble
}
\description{
The avgTrainAcc function returns the average train accuracy of an ensemble of classifiers as generated by the function \code{\link{boxEnsRBF}}.
}
\usage{
avgTrainAcc(ens)
}
\arguments{
  \item{ens}{The R object as generated by \code{\link{boxEnsRBF}}}
}
\value{
Returns a numeric value, equal to the average overall train accuracy of the ensemble. 
}
\references{
  \url{http://www.classyfire.org}
}
\seealso{
  \code{\link{getTrainAcc}},
  \code{\link{avgTrainAcc}},
  \code{\link{confMatr}}
}
\author{
Eleni Chatzimichali <hatzimihali.eleni@gmail.com>,
Conrad Bessant <c.bessant@qmul.ac.uk>
}
\examples{
data(iris)

irisClass = iris[,5]
irisData = iris[,-5]
            
ens <- boxEnsRBF(irisData, irisClass, bootNum = 10, ensNum = 20, parallel = TRUE, 
                 cpus = 4, type = "SOCK")
              
# Return the average train accuracy for the ensemble
avgTrainAcc(ens)
}
