\name{fiveNum}
\alias{fiveNum}
\title{
Five number summary
}
\description{
The fiveNum returns the "five number summary", a descriptive statistic that consists of the minimum, first (lower) quartile, median, third (upper) quartile and maximum value of a given distribution.
}
\usage{
fiveNum(permObj)
}
\arguments{
  \item{permObj}{The permutation object as generated by \code{\link{runPerm}}}
}
\value{
Returns an R object in the form of a list that contains the five number summary. The names of the returned statistics can be viewed by using the function "attributes".
}
\references{
  \url{http://www.classyfire.org}
}
\seealso{
  \link{fivenum}
}
\author{
Eleni Chatzimichali <hatzimihali.eleni@gmail.com>,
Conrad Bessant <c.bessant@qmul.ac.uk>
}
\examples{
data(iris)

irisClass = iris[,5]
irisData = iris[,-5]

permObj <- runPerm(irisData, irisClass, bootNum = 10, ensNum = 20, permNum = 5, parallel = TRUE,
                   cpus = 4, type = "SOCK")

fiveNum(permObj)
fiveNum(permObj)$median      
fiveNum(permObj)$minimum
fiveNum(permObj)$maximum
fiveNum(permObj)$upperQ
fiveNum(permObj)$lowerQ
}
