\name{fiveNumSummary}
\alias{fiveNumSummary}
\title{
Five number summary
}
\description{
The fiveNumSummary returns the "five number summary", a descriptive statistic that consists of the minimum, first (lower) quartile, median, third (upper) quartile and maximum value of a given distribution.
}
\usage{
fiveNumSummary(permObj)
}
\arguments{
  \item{permObj}{The permutation object as generated by \code{\link{runPermutation}}}
}
\value{
Returns an R object in the form of a list that contains the five number summary. The names of the returned statistics can be viewed by using the function "attributes".
}
\references{
  \url{http://www.classyfire.org}
}
\author{
Eleni Chatzimichali <hatzimihali.eleni@gmail.com>,
Conrad Bessant <c.bessant@qmul.ac.uk>
}
\examples{
data(iris)

sampleIris = iris[,5]
iris = iris[,-5]

permObj <- runPermutation(iris, sampleIris, bootNum = 10, ensembleNum = 20, permNum = 5, runParallel = TRUE, cpus = 4, type = "SOCK")

fiveNumSummary(permObj)
fiveNumSummary(permObj)$median      
fiveNumSummary(permObj)$minimum
fiveNumSummary(permObj)$maximum
fiveNumSummary(permObj)$upperQ
fiveNumSummary(permObj)$lowerQ
}
