\name{getTrainAcc}
\alias{getTrainAcc}
\title{
Train Accuracies in a classification ensemble
}
\description{
The testAcc function returns a vector of the test accuracies for all the classifiers within the ensemble as generated by \code{\link{boxEnsRBF}}.
}
\usage{
getTrainAcc(ens)
}
\arguments{
  \item{ens}{The R object as generated by \code{\link{boxEnsRBF}}}
}
\references{
  \url{http://www.classyfire.org}
}
\seealso{
  \code{\link{avgTrainAcc}},
  \code{\link{confMatr}}
}
\author{
Eleni Chatzimichali <hatzimihali.eleni@gmail.com>,
Conrad Bessant <c.bessant@qmul.ac.uk>
}
\examples{
data(iris)

irisClass = iris[,5]
irisData = iris[,-5]
            
ens <- boxEnsRBF(irisData, irisClass, bootNum = 10, ensNum = 20, parallel = TRUE, 
                 cpus = 4, type = "SOCK")
              
# Return the test accuracies for every classifier in the ensemble

getTestAcc(ens)

# Return the train accuracies for every classifier in the ensemble

getTrainAcc(ens)
}
