\name{ggClassBar}
\alias{ggClassBar}
\title{
Per class accuracies of correctly classified samples
}
\description{
The ggClassBar function generates a barplot of the per class correctly classified samples. 
}
\usage{
ggClassBar(ensObj, showText = FALSE, ...)
}
\arguments{
  \item{ensObj}{The R object as generated by \code{\link{ensRBF}}}
  \item{showText}{Boolean value, by default set to FALSE. If showText=TRUE, then the per class accuracies (\%) for all classifiers in the ensemble are displayed in the plot.}
  \item{\dots}{Further arguments and graphical parameters passed to the ggplot.}
}
\references{
  \url{http://www.classyfire.org}
}
\author{
Eleni Chatzimichali <hatzimihali.eleni@gmail.com>,
Conrad Bessant <c.bessant@qmul.ac.uk>
}
\seealso{
  \code{
    \link{ggClassPred}
  }
}
\examples{
data(iris)

sampleIris = iris[,5]
iris = iris[,-5]
            
obj <- ensRBF(iris, sampleIris, bootNum = 10, ensembleNum = 20, runParallel = TRUE, 
              cpus = 4, type = "SOCK")
              
ggClassBar(obj)
ggClassBar(obj , showText = TRUE)
}