\name{ggClassPred}
\alias{ggClassPred}
\title{
Barplot of the per class accuracies.
}
\description{
The ggClasPred function generates a barplot with the per class accuracies (\%) for all the correctly classified and misclassified samples in the classification ensemble.
}
\usage{
ggClassPred(ensObj, position = "stack", showText = FALSE, xlabel = NULL, ylabel = NULL, ...)
}
\arguments{
  \item{ensObj}{The R object as generated by \code{\link{boxEnsRBF}}}
  \item{position}{The position may be equal to either "stack" or "dodge".}
  \item{showText}{Boolean value, by default set to FALSE. If showText=TRUE, then the per class accuracies (\%) for all classifiers in the ensemble are displayed in the plot.}
  \item{xlabel}{A sub title for the x axis (optional field).}
  \item{ylabel}{A sub title for the y axis (optional field).}
  \item{\dots}{Further arguments and graphical parameters passed to the ggplot.}
}
\references{
  \url{http://www.classyfire.org}
}
\author{
Eleni Chatzimichali <hatzimihali.eleni@gmail.com>,
Conrad Bessant <c.bessant@qmul.ac.uk>
}
\seealso{
  \code{
    \link{ggClassBar}
  }
}
\examples{
data(iris)

irisClass = iris[,5]
irisData = iris[,-5]
            
ens <- boxEnsRBF(irisData, irisClass, bootNum = 10, ensNum = 20, parallel = TRUE, 
                 cpus = 4, type = "SOCK")

ggClassPred(ens, position = "dodge", showText = TRUE)
ggClassPred(ens, position = "stack")
}