\name{ggPermHist}
\alias{ggPermHist}
\title{
Permutation Histograms
}
\description{
The ggPermHist function generates a histogram of the permutation results as generated by \code{\link{runPermutation}}.
}
\usage{
ggPermHist(permObj, density = FALSE, percentiles = FALSE, mean = FALSE, median = FALSE, ...)
}
\arguments{
  \item{permObj}{The permutation object as generated by \code{\link{runPermutation}}}
  \item{density}{Boolean value, by default equal to FALSE. If density = FALSE, the histogram depicts frequencies, the counts component of the result. Instead, for density = TRUE, probability densities are plotted.}
  \item{percentiles}{Boolean value, by default equal to FALSE. If oercentiles = TRUE, the upper and lower percentiles of the distribution are depicted in the plot.}
  \item{mean}{Boolean value, by default equal to FALSE. If mean = TRUE, the mean of the distribution is depicted in the plot.}
  \item{median}{Boolean value, by default equal to FALSE. If median = TRUE, the median of the distribution is depicted in the plot.}
  \item{\dots}{Further arguments and graphical parameters passed to the ggplot.}
}
\references{
  \url{http://www.classyfire.org}
}
\author{
Eleni Chatzimichali <hatzimihali.eleni@gmail.com>,
Conrad Bessant <c.bessant@qmul.ac.uk>
}
\seealso{
  \code{
    \link{ggEnsHist}
  }
}
\examples{
data(iris)

sampleIris = iris[,5]
iris = iris[,-5]

permObj <- runPermutation(iris, sampleIris, bootNum = 10, ensembleNum = 20, permNum = 5, runParallel = TRUE, cpus = 4, type = "SOCK")

ggPermHist(permObj)
ggPermHist(permObj, density=TRUE)
ggPermHist(permObj, density=TRUE, percentiles = TRUE, mean = TRUE)
ggPermHist(permObj, density=TRUE, percentiles = TRUE, median = TRUE)
}