\name{ggplotTrend}
\alias{ggplotTrend}
\title{
Trend of the test accuracies
}
\description{
The ggplotTrend function displays the average test accuracies for every new classifier added to the ensemble, as constructed by the \code{\link{ensRBF}} function.  
}
\usage{
ggplotTrend(ensObj, xlabel = NULL, ylabel = NULL, showText = FALSE, xlims = NULL, ylims = NULL, ...)
}
\arguments{
  \item{ensObj}{The R object as generated by \code{\link{ensRBF}}}
  \item{xlabel}{A sub title for the x axis (optional field).}
  \item{ylabel}{A sub title for the y axis (optional field).}
  \item{showText}{Boolean value, by default set to FALSE. If showText=TRUE, then the values of all test accuracies in the ensemble are displayed in the plot.}
  \item{xlims}{A vector of numeric values that specifies the minimum and maximum values in the x axis.}
  \item{ylims}{A vector of numeric values that specifies the minimum and maximum values in the y axis.}
  \item{\dots}{Further arguments and graphical parameters passed to the ggplot.}
}
\references{
  \url{http://www.classyfire.org}
}
\author{
Eleni Chatzimichali <hatzimihali.eleni@gmail.com>,
Conrad Bessant <c.bessant@qmul.ac.uk>
}
\seealso{
  \code{
    \link{ggbarTest}
  }
}
\examples{
data(iris)

sampleIris = iris[,5]
iris = iris[,-5]
            
obj <- ensRBF(iris, sampleIris, bootNum = 10, ensembleNum = 20, runParallel = TRUE, 
              cpus = 4, type = "SOCK")
              
ggplotTrend(obj)
ggplotTrend(obj, showText  = TRUE)
ggplotTrend(obj, showText  = TRUE, ylims=c(90, 100))
}
