\name{overallClassPred}
\alias{overallClassPred}
\title{
Overal Class Predictions
}
\description{
The function overallClassPred returns the percentages of correctly classified samples as well as misclassified samples per given class, displayed in the form of a confusion matrix. 
}
\usage{
overallClassPred(ens)
}
\arguments{
  \item{ens}{The R object as generated by \code{\link{boxEnsRBF}}}
}
\references{
  \url{http://www.classyfire.org}
}
\seealso{
  \code{\link{confMatr}}
}
\author{
Eleni Chatzimichali <hatzimihali.eleni@gmail.com>,
Conrad Bessant <c.bessant@qmul.ac.uk>
}
\examples{
data(iris)

irisClass = iris[,5]
irisData = iris[,-5]
            
ens <- boxEnsRBF(irisData, irisClass, bootNum = 10, ensNum = 20, parallel = TRUE, 
                 cpus = 4, type = "SOCK")

overallClassPred(ens)
}
