\name{testAcc}
\alias{testAcc}
\title{
Test Accuracies (\%CC)
}
\description{
The testAcc function returns a vector of the test accuracies for all the classifiers within the ensemble as generated by \code{\link{ensRBF}}.
}
\usage{
testAcc(ens)
}
\arguments{
  \item{ens}{The R object as generated by \code{\link{ensRBF}}}
}
\references{
  \url{http://www.classyfire.org}
}
\seealso{
  \code{\link{avgTestAcc}},
  \code{\link{confMatr}}
}
\author{
Eleni Chatzimichali <hatzimihali.eleni@gmail.com>,
Conrad Bessant <c.bessant@qmul.ac.uk>
}
\examples{
data(iris)

sampleIris = iris[,5]
iris = iris[,-5]
            
obj <- ensRBF(iris, sampleIris, bootNum = 10, ensembleNum = 20, runParallel = TRUE, 
              cpus = 4, type = "SOCK")
              
# Return the test accuracies for every classifier in the ensemble

testAcc(obj)

# Return the train accuracies for every classifier in the ensemble

trainAcc(obj)
}
