\name{avgTestAcc}
\alias{avgTestAcc}
\title{
Average test accuracy for a classification ensemble
}
\description{
The avgTestAcc function returns the average test accuracy (\%CC) of an ensemble of classifiers as generated by the function \code{\link{boxEnsRBF}}.
}
\usage{
avgTestAcc(ens)
}
\arguments{
  \item{ens}{The R object (classification ensemble) as generated by \code{\link{boxEnsRBF}}}
}
\value{
Returns a numerical value, equal to the average overall test accuracy (\%CC) of the ensemble. 
}
\references{
  \url{http://www.classyfire.org}
}
\seealso{
  \code{\link{getTestAcc}},
  \code{\link{confMatr}}
}
\author{
Eleni Chatzimichali <hatzimihali.eleni@gmail.com>,
Conrad Bessant <c.bessant@qmul.ac.uk>
}
\examples{
data(iris)

irisClass = iris[,5]
irisData = iris[,-5]
            
ens <- boxEnsRBF(irisData, irisClass, bootNum = 10, ensNum = 20, parallel = TRUE, 
                 cpus = 4, type = "SOCK")

# Return the average test accuracy for the ensemble
avgTestAcc(ens)
}
