\name{avgTrainAcc}
\alias{avgTrainAcc}
\title{
Average train accuracy for a classification ensemble
}
\description{
The avgTrainAcc function returns the average train accuracy of an ensemble of classifiers as generated by the function \code{\link{ensRBF}}.
}
\usage{
avgTrainAcc(ens)
}
\arguments{
  \item{ens}{The R object as generated by \code{\link{ensRBF}}}
}
\value{
Returns a numeric value, equal to the average overall train accuracy of the ensemble. 
}
\references{
  \url{http://www.classyfire.org}
}
\seealso{
  \code{\link{trainAcc}},
  \code{\link{avgTrainAcc}},
  \code{\link{confMatr}}
}
\author{
Eleni Chatzimichali <hatzimihali.eleni@gmail.com>,
Conrad Bessant <c.bessant@qmul.ac.uk>
}
\examples{
data(iris)

sampleIris = iris[,5]
iris = iris[,-5]
            
obj <- ensRBF(iris, sampleIris, bootNum = 10, ensembleNum = 20, runParallel = TRUE, 
              cpus = 4, type = "SOCK")
# Return the average train accuracy for the ensemble

avgTrainAcc(obj)
}
