\name{confMatr}
\alias{confMatr}
\title{
Confusion matrix
}
\description{
Returns the confusion matrix for a given iteration - a given classifier within the ensemble.
}
\usage{
confMatr(iter, ens)
}
\arguments{
  \item{iter}{The number of the given iteration}
  \item{ens}{The R object as generated by \code{\link{ensRBF}}}
}
\references{
  \url{http://www.classyfire.org}
}
\seealso{
  \code{\link{overallClassPred}}
}
\author{
Eleni Chatzimichali <hatzimihali.eleni@gmail.com>,
Conrad Bessant <c.bessant@qmul.ac.uk>
}
\examples{
data(iris)

sampleIris = iris[,5]
iris = iris[,-5]
            
obj <- ensRBF(iris, sampleIris, bootNum = 10, ensembleNum = 20, runParallel = TRUE, 
              cpus = 4, type = "SOCK")

confMatr(1, obj)
}
