\name{ggEnsHist}
\alias{ggEnsHist}
\title{
Ensemble Histograms
}
\description{
The ggEnsHist function generates a histogram of the ensemble results as generated by \code{\link{boxEnsRBF}}.
}
\usage{
ggEnsHist(ensObj, density = FALSE, percentiles = FALSE, mean = FALSE, median = FALSE)
}
\arguments{
  \item{ensObj}{The ensemble object as generated by \code{\link{boxEnsRBF}}}
  \item{density}{Boolean value, by default equal to FALSE. If density = FALSE, the histogram depicts frequencies, the counts component of the result. Instead, for density = TRUE, probability densities are plotted.}
  \item{percentiles}{Boolean value, by default equal to FALSE. If oercentiles = TRUE, the upper and lower percentiles of the distribution are depicted in the plot.}
  \item{mean}{Boolean value, by default equal to FALSE. If mean = TRUE, the mean of the distribution is depicted in the plot.}
  \item{median}{Boolean value, by default equal to FALSE. If median = TRUE, the median of the distribution is depicted in the plot.}
}
\references{
  \url{http://www.classyfire.org}
}
\author{
Eleni Chatzimichali <hatzimihali.eleni@gmail.com>,
Conrad Bessant <c.bessant@qmul.ac.uk>
}
\seealso{
  \code{
    \link{ggEnsHist}
  }
}
\examples{
data(iris)

irisClass = iris[,5]
irisData = iris[,-5]
            
ens <- boxEnsRBF(irisData, irisClass, bootNum = 10, ensNum = 20, parallel = TRUE, 
                 cpus = 4, type = "SOCK")

ggEnsHist(ens)
ggEnsHist(ens, density = TRUE)
ggEnsHist(ens, density = TRUE, percentiles=TRUE)
ggEnsHist(ens, density = TRUE, percentiles=TRUE, mean=TRUE)
ggEnsHist(ens, density = TRUE, percentiles=TRUE, median=TRUE)
}

