\name{optimParams}
\alias{optimParams}
\title{
Optimal Paramaters
}
\description{
The function returns the optimal hyperparameters for the constructed classifier or classification ensemble.
}
\usage{
optimParams(ens)
}
\arguments{
  \item{ens}{The R object as generated by \code{\link{boxEnsRBF}}}
}
\value{
Returns a vector or matrix with the optimal hyperparameters as found by the optimisation process.
}
\references{
  \url{http://www.classyfire.org}
}
\author{
Eleni Chatzimichali <hatzimihali.eleni@gmail.com>,
Conrad Bessant <c.bessant@qmul.ac.uk>
}
\examples{
data(iris)

irisClass = iris[,5]
irisData = iris[,-5]
            
ens <- boxEnsRBF(irisData, irisClass, bootNum = 10, ensNum = 20, parallel = TRUE, 
                 cpus = 4, type = "SOCK")

optimParam <- optimParams(ens)
optimParam
}
