\name{overallClassPred}
\alias{overallClassPred}
\title{
Overal Class Predictions
}
\description{
The function overallClassPred returns the percentages of correctly classified samples as well as misclassified samples per given class, displayed in the form of a confusion matrix. 
}
\usage{
overallClassPred(ens)
}
\arguments{
  \item{ens}{The R object as generated by \code{\link{ensRBF}}}
}
\references{
  \url{http://www.classyfire.org}
}
\seealso{
  \code{\link{confMatr}}
}
\author{
Eleni Chatzimichali <hatzimihali.eleni@gmail.com>,
Conrad Bessant <c.bessant@qmul.ac.uk>
}
\examples{
data(iris)

sampleIris = iris[,5]
iris = iris[,-5]
            
obj <- ensRBF(iris, sampleIris, bootNum = 10, ensembleNum = 20, runParallel = TRUE, 
              cpus = 4, type = "SOCK")

overallClassPred(obj)
}
